Move sinks to core and its own namespace
