+++ lib/core/output_sink.rb
@@ -1,0 +1,13 @@
+#
+# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+# or more contributor license agreements. Licensed under the Elastic License;
+# you may not use this file except in compliance with the Elastic License.
+#
+
+# frozen_string_literal: true
+
+require 'core/output_sink/elastic_sink'
+require 'core/output_sink/console_sink'
+require 'core/output_sink/combined_sink'
+
+module Core::OutputSink ; end
